+++ pkg/sql/adaptor/mysql/adaptor_test.go
@@ -28,14 +28,26 @@ func TestOpen_Hack(t *testing.T) {
 		Expect(t, err, Failed())
 		ue := mysql.UnwrapError(err)
 		Expect(t, ue.Number, Equal(uint16(1064))) // caused by invalid database name
-
-		t.Run("Success", func(t *testing.T) {
-			d := hack.NewAdaptor(t, "mysql://root@localhost:13306/fresh")
+	})
+	t.Run("Success", func(t *testing.T) {
+		t.Run("NoOption", func(t *testing.T) {
+			dsn := "mysql://root@localhost:13306/fresh"
+			d := hack.NewAdaptor(t, dsn)
 
 			dialect := d.Dialect()
-			_, err = d.Exec(hack.Context(t), dialect.SwitchSchema("mysql"))
+			_, err := d.Exec(hack.Context(t), dialect.SwitchSchema("mysql"))
 			Expect(t, err, Succeed())
 			_, err = d.Exec(hack.Context(t), frag.Query("DROP DATABASE ?", frag.Lit("fresh")))
+			Expect(t, err, Succeed())
+		})
+		t.Run("HasOptions", func(t *testing.T) {
+			dsn := "mysql://root@localhost:13306/fresh2?interpolateParams=true"
+			d := hack.NewAdaptor(t, dsn)
+
+			dialect := d.Dialect()
+			_, err := d.Exec(hack.Context(t), dialect.SwitchSchema("mysql"))
+			Expect(t, err, Succeed())
+			_, err = d.Exec(hack.Context(t), frag.Query("DROP DATABASE ?", frag.Lit("fresh2")))
 			Expect(t, err, Succeed())
 		})
 	})