+++ pkg/sql/loggingdriver/options.go
@@ -5,9 +5,9 @@ import "database/sql/driver"
 type DriverOptionApplier func(*options)
 
 type options struct {
-	ErrorLevel          func(error) int
-	DSNParser           func(string) (string, error)
-	ValueHolderReplacer func(string, []driver.NamedValue) (string, []driver.NamedValue)
+	ErrorLevel   func(error) int
+	DSNParser    func(string) (string, error)
+	Interpolator func(string, []driver.NamedValue) (string, []driver.NamedValue)
 }
 
 func WithErrorLeveler(level func(error) int) func(*options) {
@@ -22,8 +22,8 @@ func WithDsnParser(parser func(string) (string, error)) func(*options) {
 	}
 }
 
-func WithInterpolator(replacer func(string, []driver.NamedValue) (string, []driver.NamedValue)) func(*options) {
+func WithInterpolator(interpolator func(string, []driver.NamedValue) (string, []driver.NamedValue)) func(*options) {
 	return func(o *options) {
-		o.ValueHolderReplacer = replacer
+		o.Interpolator = interpolator
 	}
 }