+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 
 <a name="HEAD"></a>
-## [HEAD](https://github.com/xoctopus/sqlx/compare/v0.3.1-rc4...HEAD)
+## [HEAD](https://github.com/xoctopus/sqlx/compare/v0.3.1-rc5...HEAD)
+
+> 0001-01-01
+
+
+<a name="v0.3.1-rc5"></a>
+## [v0.3.1-rc5](https://github.com/xoctopus/sqlx/compare/v0.3.1-rc4...v0.3.1-rc5)
 
 > 2026-06-29
 
 ### Feat
 
 * **session:** remove ExecutorX tx check in each Query and Exec
+
+### Fix
+
+* **adaptor:** interpolator for mysql
 
 
 <a name="v0.3.1-rc4"></a>